v0.1.8

+++ build/scripts-3.6/dmc_glm.py
@@ -1,13 +1,9 @@
 #!python
 """
-#=========================================================================================
+Description
+-----------
 This program performs differential CpG analysis using linear regression model based on
 beta values. 
-
-allow for covariables. 
-...
-
-#=========================================================================================
 """
 
 
@@ -26,7 +22,7 @@ __author__ = "Liguo Wang"
 __copyright__ = "Copyleft"
 __credits__ = []
 __license__ = "GPL"
-__version__="0.1.0"
+__version__="0.1.8"
 __maintainer__ = "Liguo Wang"
 __email__ = "wang.liguo@mayo.edu"
 __status__ = "Development"
@@ -36,8 +32,8 @@ def main():
 	usage="%prog [options]" + "\n"
 	parser = OptionParser(usage,version="%prog " + __version__)
 	parser.add_option("-i","--input-file",action="store",type="string",dest="input_file",help="Data file containing beta values with the 1st row containing sample IDs (must be unique) and the 1st column containing CpG positions or probe IDs (must be unique). This file can be regular text file or compressed file (*.gz, *.bz2) or accessible url.")
-	parser.add_option("-g","--group",action="store",type="string",dest="group_file",help="Group file define the biological groups of each samples as well as other covariables such as gender, age.  Sample IDs shoud match to the \"Data file\".")
-	parser.add_option("-o","--output",action="store",type='string', dest="out_file",help="Prefix of output file.")
+	parser.add_option("-g","--group",action="store",type="string",dest="group_file",help="Group file defining the biological groups of each sample as well as other covariables such as gender, age. The first varialbe is grouping variable (must be categorical), all the other variables are considered as covariates (can be categorial or continuous). Sample IDs shoud match to the \"Data file\".")
+	parser.add_option("-o","--output",action="store",type='string', dest="out_file",help="Prefix of the output file.")
 	(options,args)=parser.parse_args()
 	
 	print ()
@@ -56,16 +52,38 @@ def main():
 		parser.print_help()
 		sys.exit(103)	
 	
-	FOUT = open(options.out_file + '.pval.txt','w')
+	if not os.path.isfile(options.input_file):
+		print ("Input data file \"%s\" does not exist\n" % options.input_file) 
+		sys.exit(104)
+	if not os.path.isfile(options.group_file):
+		print ("Input group file \"%s\" does not exist\n" % options.input_file) 
+		sys.exit(105)
+	
 	ROUT = open(options.out_file + '.r','w')
 	
 	printlog("Read group file \"%s\" ..." % (options.group_file))
-	(samples,cv_names, cvs) = read_grp_file2(options.group_file)
+	(samples,cv_names, cvs, v_types) = read_grp_file2(options.group_file)
 	for cv_name in cv_names:
-		print (cv_name)
+		print ("%s: %s" % (cv_name, v_types[cv_name]))
 		for sample in samples:
-			print ('\t' + sample + '\t' + cvs[cv_name][sample])
+			print ('\t' + sample + '\t' + cvs[cv_name][sample])	
 	
+	print ('lrf1 <- function (cgid, y, %s){' % ','.join(cv_names), file=ROUT)
+	print ('try(fit <- glm(y ~ %s, family=gaussian))' % ('+'.join(cv_names)), file=ROUT)
+	print ('pvals <- coef(summary(fit))[,4]', file=ROUT)
+	print ('coefs <- coef(summary(fit))[,1]', file=ROUT)
+	print ( 'write.table(file=\"%s\",x=matrix(c(cgid, as.vector(coefs), as.vector(pvals)), nrow=1),quote=FALSE, row.names=FALSE, sep="\\t", col.names=c("ID",paste(gsub("2","",names(coefs)), "coef",sep="."), paste(gsub("2","",names(pvals)), "pval",sep=".")))' % (options.out_file + '.results.txt'),  file = ROUT) 
+	print ('}', file=ROUT)	
+	print ('\n', file=ROUT)
+
+	print ('lrf2 <- function (cgid, y,%s){' % ','.join(cv_names), file=ROUT)
+	print ('try(fit <- glm(y ~ %s, family=gaussian))' % ('+'.join(cv_names)), file=ROUT)
+	print ('pvals <- coef(summary(fit))[,4]', file=ROUT)
+	print ('coefs <- coef(summary(fit))[,1]', file=ROUT)
+	print ( 'write.table(file=\"%s\",x=matrix(c(cgid, as.vector(coefs), as.vector(pvals)), nrow=1),quote=FALSE, row.names=FALSE, sep="\\t", col.names=FALSE, append = TRUE)' % (options.out_file + '.results.txt'),  file = ROUT) 
+	print ('}', file=ROUT)	
+	print ('\n', file=ROUT)
+
 	printlog("Processing file \"%s\" ..." % (options.input_file))
 	line_num = 0
 	probe_list = []
@@ -80,6 +98,17 @@ def main():
 				if s not in sample_IDs:
 					printlog("Cannot find sample ID \"%s\" from file \"%s\"" % (s, options.input_file))
 					sys.exit(3)
+			for cv_name in cv_names:
+				if v_types[cv_name] == 'continuous':
+					print (cv_name + ' <- c(%s)' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
+				elif  v_types[cv_name] == 'categorical':
+					print (cv_name + ' <- as.factor(c(%s))' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
+				else:
+					printlog("unknown vaiable type!")
+					sys.exit(1)
+			print ('\n', file=ROUT)
+			continue
+
 			continue
 		else:
 			beta_values = []
@@ -89,55 +118,22 @@ def main():
 					beta_values.append(float(i))
 				except:
 					beta_values.append("NaN")
-			print ('',file=ROUT)
-			print ('cgid = \"%s\"' % cg_id, file=ROUT)
-			print ("y <- c(%s)" % (','.join([str(beta) for beta in beta_values])), file=ROUT)	#response variable
-			for cv_name in cv_names:
-				print (cv_name + '<- c(%s)' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
-			print ('fit <- glm(y ~ %s, family=gaussian)' % ('+'.join(cv_names)), file = ROUT)
-			print ('pval = coef(summary(fit))[,4]',file=ROUT)
-			print ('coef = coef(summary(fit))[,1]',file=ROUT)
-			print ('cat(cgid, names(pval),pval,coef, sep="\t")', file=ROUT)
-			print ('cat("\\n")', file=ROUT)
+			if line_num == 2:
+				print ('lrf1(\"%s\", c(%s), %s)' % (cg_id,  ','.join([str(i) for i in beta_values]), ','.join(cv_names)), file=ROUT)
+			else:
+				print ('lrf2(\"%s\", c(%s), %s)' % (cg_id,  ','.join([str(i) for i in beta_values]), ','.join(cv_names)), file=ROUT)
+
 	ROUT.close()
 	
 	
 	try:
 		printlog("Runing Rscript file \"%s\" ..." % (options.out_file + '.r'))
-		subprocess.call("Rscript %s >%s 2>%s" % (options.out_file + '.r', options.out_file + '.r.results.txt',options.out_file + '.r.warnings.txt' ), shell=True)
+		subprocess.call("Rscript %s 2>%s" % (options.out_file + '.r', options.out_file + '.warnings.txt' ), shell=True)
 	except:
 		print ("Error: cannot run Rscript: \"%s\"" % (options.out_file + '.r'), file=sys.stderr)
 		sys.exit(1)
 	
-	
-	printlog("Reading file \"%s\" ..." % (options.out_file + '.r.results.txt'))
-	glm_results = {}
-	for l in open(options.out_file + '.r.results.txt'):
-		l = l.strip()
-		l = l.replace(')','')
-		l = l.replace('(','')
-		f = l.split('\t')
-		cgID = f[0]
-		tmp = f[1:]
-		chunk_size = int(len(tmp)/3)
-		sub_lists = [tmp[i:i+chunk_size] for i in range(0,len(tmp),chunk_size)]
-		v_names = sub_lists[0][1:]
-		v_pvals = sub_lists[1][1:]
-		v_coefs = sub_lists[2][1:]
-		glm_results[cgID] = [v_pvals, v_coefs]
-	
-	printlog("Results saved to \"%s\" ..." % (options.out_file + '.pval.txt'))
-	line_num = 0
-	for l in ireader.reader(options.input_file):
-		line_num += 1
-		f = l.split()
-		if line_num == 1:
-			print (l + '\t' + '\t'.join([i + '.pval' for i in v_names]) + '\t' + '\t'.join([i + '.coef' for i in v_names]), file=FOUT)
-		else:
-			cgID = f[0]
-			print (l + '\t' + '\t'.join(glm_results[cgID][0]) + '\t' + '\t'.join(glm_results[cgID][1]), file=FOUT)
-	
-	FOUT.close()
 
+	
 if __name__=='__main__':
 	main()
