+++ build/scripts-3.6/dmc_logit.py
@@ -1,20 +1,21 @@
 #!python
 """
-#=========================================================================================
+Description
+-----------
 This program performs differential CpG analysis using logistic regression model based on
 methylation proportions (in the form of "c,n", where "c" indicates "Number of reads with
 methylated C", and "n" indicates "Number of total reads". Both c and n are  non-negative
-integers and c <= n). Below example showing input data on 2 CpGs of 3 groups (A,B, and C)
+integers and c <= n). 
+
+Example of input data
+---------------------
+Below example showing input data on 2 CpGs of 3 groups (A,B, and C)
 with each group has 3 replicates:
  
 cgID  A_1   A_2   A_3   B_1   B_2   B_3   C_1   C_2   C_3
 CpG_1 129,170 166,178 7,9 1 6,16  10,10 10,15 11,15 16,22 20,36     
 CpG_2 0,77  0,99  0,85  0,77  1,37  3,37  0,42  0,153 0,6
 
-allow for covariables. 
-...
-
-#=========================================================================================
 """
 
 
@@ -34,7 +35,7 @@ __author__ = "Liguo Wang"
 __copyright__ = "Copyleft"
 __credits__ = []
 __license__ = "GPL"
-__version__="0.1.0"
+__version__="0.1.8"
 __maintainer__ = "Liguo Wang"
 __email__ = "wang.liguo@mayo.edu"
 __status__ = "Development"
@@ -44,8 +45,9 @@ def main():
 	usage="%prog [options]" + "\n"
 	parser = OptionParser(usage,version="%prog " + __version__)
 	parser.add_option("-i","--input-file",action="store",type="string",dest="input_file",help="Data file containing methylation proportions (represented by \"methyl_count,total_count\", eg. \"20,30\") with the 1st row containing sample IDs (must be unique) and the 1st column containing CpG positions or probe IDs (must be unique). This file can be a regular text file or compressed file (*.gz, *.bz2) or accessible url..")
-	parser.add_option("-g","--group",action="store",type="string",dest="group_file",help="Group file define the biological groups of each samples as well as other covariables such as gender, age.  Sample IDs shoud match to the \"Data file\".")
-	parser.add_option("-o","--output",action="store",type='string', dest="out_file",help="Prefix of output file.")
+	parser.add_option("-g","--group",action="store",type="string",dest="group_file",help="Group file defining the biological groups of each sample as well as other covariables such as gender, age. The first varialbe is grouping variable (must be categorical), all the other variables are considered as covariates (can be categorial or continuous). Sample IDs shoud match to the \"Data file\".")
+	parser.add_option("-f","--family",action="store",type="int",dest="family_func",default=1, help="Error distribution and link function to be used in the GLM model. Can be integer 1 or 2 with 1 = \"quasibinomial\" and 2 = \"binomial\". Default=%default.")
+	parser.add_option("-o","--output",action="store",type='string', dest="out_file",help="Prefix of the output file.")
 	(options,args)=parser.parse_args()
 	
 	print ()
@@ -63,17 +65,46 @@ def main():
 		print (__doc__)
 		parser.print_help()
 		sys.exit(103)	
+	if not os.path.isfile(options.input_file):
+		print ("Input data file \"%s\" does not exist\n" % options.input_file) 
+		sys.exit(104)
+	if not os.path.isfile(options.group_file):
+		print ("Input group file \"%s\" does not exist\n" % options.input_file) 
+		sys.exit(105)
 	
-	FOUT = open(options.out_file + '.pval.txt','w')
 	ROUT = open(options.out_file + '.r','w')
-	
+	family = {1:'quasibinomial', 2:'binomial',}
+	if not options.family_func in family.keys():
+		print ("Incorrect value of '-f'!") 
+		sys.exit(106)
+		
 	printlog("Read group file \"%s\" ..." % (options.group_file))
-	(samples,cv_names, cvs) = read_grp_file2(options.group_file)
+	(samples,cv_names, cvs, v_types) = read_grp_file2(options.group_file)
 	for cv_name in cv_names:
-		print (cv_name)
+		print ("%s: %s" % (cv_name, v_types[cv_name]))
 		for sample in samples:
 			print ('\t' + sample + '\t' + cvs[cv_name][sample])
 	
+	print ('lrf1 <- function (cgid, m,t,%s){' % ','.join(cv_names), file=ROUT)
+	print ('try(fit <- glm(cbind(m,t - m) ~ %s, family=%s))' % ('+'.join(cv_names),family[options.family_func]), file=ROUT)
+	print ('pvals <- coef(summary(fit))[,4]', file=ROUT)
+	print ('coefs <- coef(summary(fit))[,1]', file=ROUT)
+	print ('\tif(max(pvals, na.rm=T)>1){pvals = pvals + NA}', file=ROUT)
+	print ('\tif(sum(m, na.rm=T) == 0){pvals = pvals + NA}', file=ROUT)
+	print ( 'write.table(file=\"%s\",x=matrix(c(cgid, as.vector(coefs), as.vector(pvals)), nrow=1),quote=FALSE, row.names=FALSE, sep="\\t", col.names=c("ID",paste(gsub("2","",names(coefs)), "coef",sep="."), paste(gsub("2","",names(pvals)), "pval",sep=".")))' % (options.out_file + '.results.txt'),  file = ROUT) 
+	print ('}', file=ROUT)	
+	print ('\n', file=ROUT)
+
+	print ('lrf2 <- function (cgid, m,t,%s){' % ','.join(cv_names), file=ROUT)
+	print ('try(fit <- glm(cbind(m,t - m) ~ %s, family=%s))' % ('+'.join(cv_names),family[options.family_func]), file=ROUT)
+	print ('pvals <- coef(summary(fit))[,4]', file=ROUT)
+	print ('coefs <- coef(summary(fit))[,1]', file=ROUT)
+	print ('\tif(max(pvals, na.rm=T)>1){pvals = pvals + NA}', file=ROUT)
+	print ('\tif(sum(m, na.rm=T) == 0){pvals = pvals + NA}', file=ROUT)
+	print ( 'write.table(file=\"%s\",x=matrix(c(cgid, as.vector(coefs), as.vector(pvals)), nrow=1),quote=FALSE, row.names=FALSE, sep="\\t", col.names=FALSE, append = TRUE)' % (options.out_file + '.results.txt'),  file = ROUT) 
+	print ('}', file=ROUT)	
+	print ('\n', file=ROUT)
+		
 	printlog("Processing file \"%s\" ..." % (options.input_file))
 	line_num = 0
 	probe_list = []
@@ -88,6 +119,18 @@ def main():
 				if s not in sample_IDs:
 					printlog("Cannot find sample ID \"%s\" from file \"%s\"" % (s, options.input_file))
 					sys.exit(3)
+			#for cv_name in cv_names:
+			#	print (cv_name + ' <- c(%s)' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
+			for cv_name in cv_names:
+				if v_types[cv_name] == 'continuous':
+					print (cv_name + ' <- c(%s)' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
+				elif  v_types[cv_name] == 'categorical':
+					print (cv_name + ' <- as.factor(c(%s))' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
+				else:
+					printlog("unknown vaiable type!")
+					sys.exit(1)
+
+			print ('\n', file=ROUT)
 			continue
 		else:
 			methyl_reads = []			# c
@@ -109,57 +152,22 @@ def main():
 					else:
 						printlog("Incorrect data format!")
 						print (f)
-						sys.exit(1)						
-			print ('',file=ROUT)
-			print ('cgid = \"%s\"' % cg_id, file=ROUT)
-			print ("y <- c(%s)" % (','.join([str(read) for read in methyl_reads])), file=ROUT)	#response variable
-			print ("total_reads <- c(%s)" % (','.join([str(read) for read in total_reads])), file=ROUT)	#For a binomial GLM prior weights are used to give the number of trials when the response is the proportion of successes
-			for cv_name in cv_names:
-				print (cv_name + '<- c(%s)' % (','.join([str(cvs[cv_name][s]) for s in  sample_IDs  ])), file = ROUT)
-			print ('fit <- glm(y/total_reads ~ %s, weights=total_reads, family=binomial)' % ('+'.join(cv_names)), file = ROUT)
-			print ('pval = coef(summary(fit))[,4]',file=ROUT)
-			print ('coef = coef(summary(fit))[,1]',file=ROUT)
-			print ('cat(cgid, names(pval),pval,coef, sep="\t")', file=ROUT)
-			print ('cat("\\n")', file=ROUT)
+						sys.exit(1)		
+			if line_num == 2:
+				print ('lrf1(\"%s\", c(%s), c(%s), %s)' % (cg_id, ','.join([str(read) for read in methyl_reads]), ','.join([str(read) for read in total_reads]), ','.join(cv_names)), file=ROUT)
+			else:
+				print ('lrf2(\"%s\", c(%s), c(%s), %s)' % (cg_id, ','.join([str(read) for read in methyl_reads]), ','.join([str(read) for read in total_reads]), ','.join(cv_names)), file=ROUT)
+
 	ROUT.close()
 	
 	
 	try:
 		printlog("Runing Rscript file \"%s\" ..." % (options.out_file + '.r'))
-		subprocess.call("Rscript %s >%s 2>%s" % (options.out_file + '.r', options.out_file + '.r.results.txt',options.out_file + '.r.warnings.txt' ), shell=True)
+		subprocess.call("Rscript %s 2>%s" % (options.out_file + '.r', options.out_file + '.warnings.txt' ), shell=True)
 	except:
 		print ("Error: cannot run Rscript: \"%s\"" % (options.out_file + '.r'), file=sys.stderr)
 		sys.exit(1)
 	
-	
-	printlog("Reading file \"%s\" ..." % (options.out_file + '.r.results.txt'))
-	glm_results = {}
-	for l in open(options.out_file + '.r.results.txt'):
-		l = l.strip()
-		l = l.replace(')','')
-		l = l.replace('(','')
-		f = l.split('\t')
-		cgID = f[0]
-		tmp = f[1:]
-		chunk_size = int(len(tmp)/3)
-		sub_lists = [tmp[i:i+chunk_size] for i in range(0,len(tmp),chunk_size)]
-		v_names = sub_lists[0][1:]
-		v_pvals = sub_lists[1][1:]
-		v_coefs = sub_lists[2][1:]
-		glm_results[cgID] = [v_pvals, v_coefs]
-	
-	printlog("Results saved to \"%s\" ..." % (options.out_file + '.pval.txt'))
-	line_num = 0
-	for l in ireader.reader(options.input_file):
-		line_num += 1
-		f = l.split()
-		if line_num == 1:
-			print (l + '\t' + '\t'.join([i + '.pval' for i in v_names]) + '\t' + '\t'.join([i + '.coef' for i in v_names]), file=FOUT)
-		else:
-			cgID = f[0]
-			print (l + '\t' + '\t'.join(glm_results[cgID][0]) + '\t' + '\t'.join(glm_results[cgID][1]), file=FOUT)
-	
-	FOUT.close()
-
 if __name__=='__main__':
 	main()
+
